// src/app/domains/customer-management/domain/models/customer.model.ts
/**
 *
 * Interface that represents a Customer in the system.
 *
 * Contains the basic contact information of customers
 * associated with a workshop.
 *
 * @interface
 * @model
 *
 */

export interface Customer {
  /**
 * Unique ID of the customer (generated by json-server)
 */
  id?: string|number;
  /**
 * ID of the workshop to which this customer belongs
 */
  workshopId?: string;
  /**
 * Full name of the customer
 */
  fullName: string;
  /**
 * National Identity Document (DNI)
 */
  dni: string;
  /**
 * Email address of the customer
 */
  email: string;
  /**
 * Contact phone number of the customer
 */
  phone: string;
}